Move SubmapController to mapping.
